feat : Configured Project For serverless hosting on Vercel
