--- src/main.ts
@@ -1,48 +1,0 @@
-import { NestFactory } from '@nestjs/core';
-import { ValidationPipe } from '@nestjs/common';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-import { AppModule } from './app.module';
-
-async function bootstrap() {
-  const app = await NestFactory.create(AppModule);
-  app.useGlobalPipes(
-    new ValidationPipe({
-      transform: true,
-      whitelist: true,
-      forbidNonWhitelisted: true,
-    }),
-  );
-  app.enableCors();
-
-  // Swagger configuration
-  const config = new DocumentBuilder()
-    .setTitle('Fortrez API')
-    .setDescription('The Fortrez crowdfunding platform API documentation')
-    .setVersion('1.0')
-    .addBearerAuth(
-      {
-        type: 'http',
-        scheme: 'bearer',
-        bearerFormat: 'JWT',
-        name: 'JWT',
-        description: 'Enter JWT token',
-        in: 'header',
-      },
-      'JWT-auth', // This name is used as the key for the security requirement in the @ApiBearerAuth() decorator
-    )
-    .build();
-  
-  const document = SwaggerModule.createDocument(app, config);
-  SwaggerModule.setup('api', app, document, {
-    swaggerOptions: {
-      persistAuthorization: true,
-    },
-  });
-
-  await app.listen(process.env.PORT ?? 3000);
-  console.log(`Application is running on port ${process.env.PORT}`);
-}
-void bootstrap().catch((err) => {
-  console.error('Failed to start application:', err);
-  process.exit(1);
-});
