+++ vercel.json
@@ -1,0 +1,5 @@
+{
+  "version": 2,
+  "builds": [{ "src": "src/vercel.ts", "use": "@vercel/node" }],
+  "routes": [{ "src": "/(.*)", "dest": "src/vercel.ts" }]
+}
